Modularize codebase into 3-file structure
- Create utils.py with helper functions for preprocessing and correlation
  - _preprocess_expression(): log2 + min-max normalization
  - _load_data(): load raw data from files
  - _extract_data(): extract from AnnData/DataFrame
  - _create_protein_pairs(): correlation calculation
  - _pairs_after_cutoff(): filtering logic

- Create models.py with VAE class
  - Move VAE model class to separate module

- Refactor fava.py to use new modules
  - Remove moved functions/classes
  - Add imports from utils and models
  - Keep cook(), main(), and argument_parser()
  - Reduced from ~533 to ~235 lines

- Update __init__.py to use relative imports

Benefits:
- Better code organization and separation of concerns
- Easier to test and maintain individual components
- Public API remains unchanged (backward compatible)
