+++ src/favapy/__init__.py
@@ -1,4 +1,4 @@
-from favapy.fava import cook, main
+from .fava import cook, main
 
 __version__ = "1.0.1"
 __all__ = ["cook", "main"]
